requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.16
albumentations==1.4.15
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
cachetools==5.5.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cookiecutter==2.6.0
deprecation==2.1.0
easydict==1.13
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-limiter==0.1.6
filelock==3.16.1
flatbuffers==24.3.25
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gotrue==2.9.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.2
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==23.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
postgrest==0.17.1
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.2.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
realtime==2.0.5
redis==5.1.1
referencing==0.35.1
rembg==2.0.59
repath==0.9.0
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.40.0
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
sympy==1.13.3
text-unidecode==1.3
tifffile==2024.9.20
timm==1.0.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transparent-background==1.3.3
triton==3.0.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.1
wget==3.2
yarl==1.15.2
